Working on creating multiple foods
Work on creating multiple foods so you don't chase across the level to
have it disappear on you.

It was suggested that the food shouldn't move, and I thought making
multiple foods might appeal to both player suggestions, and my own
feelings about the game.

Couple with the ability to travel through the level... games of snake
should end up being pretty quick to start.